made reloads work
no longer need  to restart modo when devving
